refactor(domainowned): mv components
